requirements.txt
Django==4.2.0
djangorestframework==3.14.0
pandas==2.0.0
matplotlib==3.7.1
seaborn==0.12.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.5.0
scipy==1.10.1
numpy==1.24.3
